Renamed metrics module to metric_events
