+++ corehq/apps/enterprise/api/resources.py
@@ -22,7 +22,7 @@ from corehq.apps.api.resources.auth import ODataAuthentication
 from corehq.apps.api.resources.meta import get_hq_throttle
 from corehq.apps.api.keyset_paginator import KeysetPaginator
 from corehq.apps.enterprise.enterprise import EnterpriseReport
-from corehq.apps.enterprise.metrics import ENTERPRISE_API_ACCESS
+from corehq.apps.enterprise.metric_events import ENTERPRISE_API_ACCESS
 from corehq.apps.enterprise.iterators import IterableEnterpriseFormQuery, EnterpriseFormReportConverter
 
 from corehq.apps.enterprise.tasks import generate_enterprise_report, ReportTaskProgress
