+++ corehq/apps/enterprise/views.py
@@ -26,7 +26,7 @@ from corehq.apps.accounting.decorators import always_allow_project_access
 from corehq.apps.analytics.tasks import record_event
 from corehq.apps.enterprise.decorators import require_enterprise_admin
 from corehq.apps.enterprise.exceptions import TooMuchRequestedDataError
-from corehq.apps.enterprise.metrics import ENTERPRISE_REPORT_REQUEST
+from corehq.apps.enterprise.metric_events import ENTERPRISE_REPORT_REQUEST
 from corehq.apps.enterprise.mixins import ManageMobileWorkersMixin
 from corehq.apps.enterprise.models import EnterprisePermissions
 from corehq.apps.enterprise.tasks import clear_enterprise_permissions_cache_for_all_users
